% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{detect_data_types}
\alias{detect_data_types}
\title{Detect data types in a dataframe}
\usage{
detect_data_types(data)
}
\arguments{
\item{data}{A dataframe to analyze}
}
\value{
A list with categorized column names
}
\description{
Detect data types in a dataframe
}
\examples{
detect_data_types(iris)
}
