% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{print.data_summary}
\alias{print.data_summary}
\title{Print method for data_summary objects}
\usage{
\method{print}{data_summary}(x, ...)
}
\arguments{
\item{x}{A data_summary object}

\item{...}{Additional arguments}
}
\description{
Print method for data_summary objects
}
