% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_analysis.R
\name{print.missing_analysis}
\alias{print.missing_analysis}
\title{Print method for missing_analysis objects}
\usage{
\method{print}{missing_analysis}(x, ...)
}
\arguments{
\item{x}{A missing_analysis object}

\item{...}{Additional arguments}
}
\description{
Print method for missing_analysis objects
}
