% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_recommendations.R
\name{print.model_recommendations}
\alias{print.model_recommendations}
\title{Print method for model_recommendations objects}
\usage{
\method{print}{model_recommendations}(x, ...)
}
\arguments{
\item{x}{A model_recommendations object}

\item{...}{Additional arguments}
}
\description{
Print method for model_recommendations objects
}
