% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{print.outlier_detection}
\alias{print.outlier_detection}
\title{Print method for outlier_detection objects}
\usage{
\method{print}{outlier_detection}(x, ...)
}
\arguments{
\item{x}{An outlier_detection object}

\item{...}{Additional arguments}
}
\description{
Print method for outlier_detection objects
}
