% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eda_app.R
\name{run_eda_app}
\alias{run_eda_app}
\title{Launch the EDAhelper Shiny App}
\usage{
run_eda_app(browser = FALSE, ...)
}
\arguments{
\item{browser}{Logical, whether to launch the app in a browser (TRUE) or in the RStudio viewer (FALSE).
Default is FALSE.}

\item{...}{Additional arguments to pass to \code{shiny::runApp()}.}
}
\value{
No return value, called for side effects.
}
\description{
This function launches an interactive Shiny application that provides a graphical
user interface for the EDAhelper package. The app allows users to upload data,
perform exploratory data analysis, visualize distributions with subgroup analysis,
detect outliers, and get model recommendations.
}
\examples{
\dontrun{
# Launch the EDAhelper Shiny app
run_eda_app()

# Launch in browser
run_eda_app(browser = TRUE)
}

}
